config_Get finally works...
