+++ test/integration/database_session_9_test.rb
@@ -25,11 +25,6 @@ class TestDatabaseSession < MiniTest::Unit::TestCase
 
     refute @connection.closed?
   end
-  
-
-  def test_config_get
-       @session.config_get(@session, "cache")
-  end
 
   def test_query
     result = @session.query("SELECT FROM OUser")
